Renamed model.dll to Schemazen.Library.dll
