+++ model/Models/BatchSqlParser.cs
@@ -2,7 +2,7 @@
 using System.Linq;
 using System.Text.RegularExpressions;
 
-namespace SchemaZen.model {
+namespace SchemaZen.Library.Models {
 	internal enum State {
 		Searching,
 		InOneLineComment,
