+++ model/Models/ConstraintColumn.cs
@@ -1,4 +1,4 @@
-﻿namespace SchemaZen.model
+﻿namespace SchemaZen.Library.Models
 {
     public class ConstraintColumn
     {
