+++ model/Models/Role.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace SchemaZen.model
+﻿namespace SchemaZen.Library.Models
 {
     public class Role : IScriptable, INameable
     {
