+++ test/BatchSqlParserTester.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 using NUnit.Framework;
-using SchemaZen.model;
+using SchemaZen.Library.Models;
 
-namespace SchemaZen.test {
+namespace SchemaZen.Tests {
 	[TestFixture]
 	public class BatchSqlParserTester {
 		[Test]
