+++ test/TestHelper.cs
@@ -1,9 +1,9 @@
 ﻿using System;
 using System.Data.SqlClient;
 using NUnit.Framework;
-using SchemaZen.model;
+using SchemaZen.Library.Models;
 
-namespace SchemaZen.test {
+namespace SchemaZen.Tests {
 	[SetUpFixture]
 	public class TestHelper {
 		public static bool EchoSql {
